third_party/blink/web_tests/platform/webview/external/wpt/webusb/idlharness.https.any.worker-expected.txt

This is a testharness.js-based test.
Found 255 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] USB device setup
  assert_implements: missing navigator.usb undefined
[FAIL] USB interface: existence and properties of interface object
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface object length
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface object name
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface: attribute onconnect
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface: attribute ondisconnect
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface: operation getDevices()
  assert_own_property: self does not have own property "USB" expected property "USB" missing
[FAIL] USB interface: member requestDevice
  Cannot use 'in' operator to search for 'requestDevice' in undefined
[FAIL] USB must be primary interface of navigator.usb
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of navigator.usb
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] USB interface: navigator.usb must inherit property "onconnect" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] USB interface: navigator.usb must inherit property "ondisconnect" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] USB interface: navigator.usb must inherit property "getDevices()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] USB interface: navigator.usb must not have property "requestDevice"
  Cannot use 'in' operator to search for 'requestDevice' in undefined
[FAIL] USBConnectionEvent interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
[FAIL] USBConnectionEvent interface object length
  assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
[FAIL] USBConnectionEvent interface object name
  assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
[FAIL] USBConnectionEvent interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
[FAIL] USBConnectionEvent interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
[FAIL] USBConnectionEvent interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
[FAIL] USBConnectionEvent interface: attribute device
  assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
[FAIL] USBConnectionEvent must be primary interface of usbConnectionEvent
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbConnectionEvent is not defined"
[FAIL] Stringification of usbConnectionEvent
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbConnectionEvent is not defined"
[FAIL] USBConnectionEvent interface: usbConnectionEvent must inherit property "device" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbConnectionEvent is not defined"
[FAIL] USBInTransferResult interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
[FAIL] USBInTransferResult interface object length
  assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
[FAIL] USBInTransferResult interface object name
  assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
[FAIL] USBInTransferResult interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
[FAIL] USBInTransferResult interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
[FAIL] USBInTransferResult interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
[FAIL] USBInTransferResult interface: attribute data
  assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
[FAIL] USBInTransferResult interface: attribute status
  assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
[FAIL] USBInTransferResult must be primary interface of new USBInTransferResult("ok")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBInTransferResult is not defined"
[FAIL] Stringification of new USBInTransferResult("ok")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBInTransferResult is not defined"
[FAIL] USBInTransferResult interface: new USBInTransferResult("ok") must inherit property "data" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBInTransferResult is not defined"
[FAIL] USBInTransferResult interface: new USBInTransferResult("ok") must inherit property "status" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBInTransferResult is not defined"
[FAIL] USBOutTransferResult interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
[FAIL] USBOutTransferResult interface object length
  assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
[FAIL] USBOutTransferResult interface object name
  assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
[FAIL] USBOutTransferResult interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
[FAIL] USBOutTransferResult interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
[FAIL] USBOutTransferResult interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
[FAIL] USBOutTransferResult interface: attribute bytesWritten
  assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
[FAIL] USBOutTransferResult interface: attribute status
  assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
[FAIL] USBOutTransferResult must be primary interface of new USBOutTransferResult("ok")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBOutTransferResult is not defined"
[FAIL] Stringification of new USBOutTransferResult("ok")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBOutTransferResult is not defined"
[FAIL] USBOutTransferResult interface: new USBOutTransferResult("ok") must inherit property "bytesWritten" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBOutTransferResult is not defined"
[FAIL] USBOutTransferResult interface: new USBOutTransferResult("ok") must inherit property "status" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBOutTransferResult is not defined"
[FAIL] USBIsochronousInTransferPacket interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
[FAIL] USBIsochronousInTransferPacket interface object length
  assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
[FAIL] USBIsochronousInTransferPacket interface object name
  assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
[FAIL] USBIsochronousInTransferPacket interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
[FAIL] USBIsochronousInTransferPacket interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
[FAIL] USBIsochronousInTransferPacket interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
[FAIL] USBIsochronousInTransferPacket interface: attribute data
  assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
[FAIL] USBIsochronousInTransferPacket interface: attribute status
  assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
[FAIL] USBIsochronousInTransferPacket must be primary interface of new USBIsochronousInTransferPacket("ok")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferPacket is not defined"
[FAIL] Stringification of new USBIsochronousInTransferPacket("ok")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferPacket is not defined"
[FAIL] USBIsochronousInTransferPacket interface: new USBIsochronousInTransferPacket("ok") must inherit property "data" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferPacket is not defined"
[FAIL] USBIsochronousInTransferPacket interface: new USBIsochronousInTransferPacket("ok") must inherit property "status" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferPacket is not defined"
[FAIL] USBIsochronousInTransferResult interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
[FAIL] USBIsochronousInTransferResult interface object length
  assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
[FAIL] USBIsochronousInTransferResult interface object name
  assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
[FAIL] USBIsochronousInTransferResult interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
[FAIL] USBIsochronousInTransferResult interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
[FAIL] USBIsochronousInTransferResult interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
[FAIL] USBIsochronousInTransferResult interface: attribute data
  assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
[FAIL] USBIsochronousInTransferResult interface: attribute packets
  assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
[FAIL] USBIsochronousInTransferResult must be primary interface of new USBIsochronousInTransferResult([])
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferResult is not defined"
[FAIL] Stringification of new USBIsochronousInTransferResult([])
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferResult is not defined"
[FAIL] USBIsochronousInTransferResult interface: new USBIsochronousInTransferResult([]) must inherit property "data" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferResult is not defined"
[FAIL] USBIsochronousInTransferResult interface: new USBIsochronousInTransferResult([]) must inherit property "packets" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferResult is not defined"
[FAIL] USBIsochronousOutTransferPacket interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
[FAIL] USBIsochronousOutTransferPacket interface object length
  assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
[FAIL] USBIsochronousOutTransferPacket interface object name
  assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
[FAIL] USBIsochronousOutTransferPacket interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
[FAIL] USBIsochronousOutTransferPacket interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
[FAIL] USBIsochronousOutTransferPacket interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
[FAIL] USBIsochronousOutTransferPacket interface: attribute bytesWritten
  assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
[FAIL] USBIsochronousOutTransferPacket interface: attribute status
  assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
[FAIL] USBIsochronousOutTransferPacket must be primary interface of new USBIsochronousOutTransferPacket("ok")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferPacket is not defined"
[FAIL] Stringification of new USBIsochronousOutTransferPacket("ok")
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferPacket is not defined"
[FAIL] USBIsochronousOutTransferPacket interface: new USBIsochronousOutTransferPacket("ok") must inherit property "bytesWritten" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferPacket is not defined"
[FAIL] USBIsochronousOutTransferPacket interface: new USBIsochronousOutTransferPacket("ok") must inherit property "status" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferPacket is not defined"
[FAIL] USBIsochronousOutTransferResult interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
[FAIL] USBIsochronousOutTransferResult interface object length
  assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
[FAIL] USBIsochronousOutTransferResult interface object name
  assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
[FAIL] USBIsochronousOutTransferResult interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
[FAIL] USBIsochronousOutTransferResult interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
[FAIL] USBIsochronousOutTransferResult interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
[FAIL] USBIsochronousOutTransferResult interface: attribute packets
  assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
[FAIL] USBIsochronousOutTransferResult must be primary interface of new USBIsochronousOutTransferResult([])
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferResult is not defined"
[FAIL] Stringification of new USBIsochronousOutTransferResult([])
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferResult is not defined"
[FAIL] USBIsochronousOutTransferResult interface: new USBIsochronousOutTransferResult([]) must inherit property "packets" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferResult is not defined"
[FAIL] USBDevice interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface object length
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface object name
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute usbVersionMajor
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute usbVersionMinor
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute usbVersionSubminor
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute deviceClass
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute deviceSubclass
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute deviceProtocol
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute vendorId
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute productId
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute deviceVersionMajor
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute deviceVersionMinor
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute deviceVersionSubminor
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute manufacturerName
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute productName
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute serialNumber
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute configuration
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute configurations
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: attribute opened
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation open()
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation close()
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation forget()
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation selectConfiguration(octet)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation claimInterface(octet)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation releaseInterface(octet)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation selectAlternateInterface(octet, octet)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation controlTransferIn(USBControlTransferParameters, unsigned short)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation controlTransferOut(USBControlTransferParameters, optional BufferSource)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation clearHalt(USBDirection, octet)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation transferIn(octet, unsigned long)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation transferOut(octet, BufferSource)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation isochronousTransferIn(octet, sequence<unsigned long>)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation isochronousTransferOut(octet, BufferSource, sequence<unsigned long>)
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice interface: operation reset()
  assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
[FAIL] USBDevice must be primary interface of usbDevice
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] Stringification of usbDevice
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "usbVersionMajor" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "usbVersionMinor" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "usbVersionSubminor" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "deviceClass" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "deviceSubclass" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "deviceProtocol" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "vendorId" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "productId" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "deviceVersionMajor" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "deviceVersionMinor" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "deviceVersionSubminor" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "manufacturerName" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "productName" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "serialNumber" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "configuration" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "configurations" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "opened" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "open()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "close()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "forget()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "selectConfiguration(octet)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling selectConfiguration(octet) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "claimInterface(octet)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling claimInterface(octet) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "releaseInterface(octet)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling releaseInterface(octet) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "selectAlternateInterface(octet, octet)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling selectAlternateInterface(octet, octet) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "controlTransferIn(USBControlTransferParameters, unsigned short)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling controlTransferIn(USBControlTransferParameters, unsigned short) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "controlTransferOut(USBControlTransferParameters, optional BufferSource)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling controlTransferOut(USBControlTransferParameters, optional BufferSource) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "clearHalt(USBDirection, octet)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling clearHalt(USBDirection, octet) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "transferIn(octet, unsigned long)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling transferIn(octet, unsigned long) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "transferOut(octet, BufferSource)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling transferOut(octet, BufferSource) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "isochronousTransferIn(octet, sequence<unsigned long>)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling isochronousTransferIn(octet, sequence<unsigned long>) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "isochronousTransferOut(octet, BufferSource, sequence<unsigned long>)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: calling isochronousTransferOut(octet, BufferSource, sequence<unsigned long>) on usbDevice with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBDevice interface: usbDevice must inherit property "reset()" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbDevice is not defined"
[FAIL] USBConfiguration interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration interface object length
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration interface object name
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration interface: attribute configurationValue
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration interface: attribute configurationName
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration interface: attribute interfaces
  assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
[FAIL] USBConfiguration must be primary interface of usbConfiguration
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbConfiguration is not defined"
[FAIL] Stringification of usbConfiguration
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbConfiguration is not defined"
[FAIL] USBConfiguration interface: usbConfiguration must inherit property "configurationValue" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbConfiguration is not defined"
[FAIL] USBConfiguration interface: usbConfiguration must inherit property "configurationName" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbConfiguration is not defined"
[FAIL] USBConfiguration interface: usbConfiguration must inherit property "interfaces" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbConfiguration is not defined"
[FAIL] USBInterface interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface object length
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface object name
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface: attribute interfaceNumber
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface: attribute alternate
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface: attribute alternates
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface interface: attribute claimed
  assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
[FAIL] USBInterface must be primary interface of usbInterface
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbInterface is not defined"
[FAIL] Stringification of usbInterface
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbInterface is not defined"
[FAIL] USBInterface interface: usbInterface must inherit property "interfaceNumber" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbInterface is not defined"
[FAIL] USBInterface interface: usbInterface must inherit property "alternate" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbInterface is not defined"
[FAIL] USBInterface interface: usbInterface must inherit property "alternates" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbInterface is not defined"
[FAIL] USBInterface interface: usbInterface must inherit property "claimed" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbInterface is not defined"
[FAIL] USBAlternateInterface interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface object length
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface object name
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: attribute alternateSetting
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: attribute interfaceClass
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: attribute interfaceSubclass
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: attribute interfaceProtocol
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: attribute interfaceName
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface interface: attribute endpoints
  assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
[FAIL] USBAlternateInterface must be primary interface of usbAlternateInterface
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbAlternateInterface is not defined"
[FAIL] Stringification of usbAlternateInterface
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbAlternateInterface is not defined"
[FAIL] USBAlternateInterface interface: usbAlternateInterface must inherit property "alternateSetting" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbAlternateInterface is not defined"
[FAIL] USBAlternateInterface interface: usbAlternateInterface must inherit property "interfaceClass" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbAlternateInterface is not defined"
[FAIL] USBAlternateInterface interface: usbAlternateInterface must inherit property "interfaceSubclass" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbAlternateInterface is not defined"
[FAIL] USBAlternateInterface interface: usbAlternateInterface must inherit property "interfaceProtocol" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbAlternateInterface is not defined"
[FAIL] USBAlternateInterface interface: usbAlternateInterface must inherit property "interfaceName" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbAlternateInterface is not defined"
[FAIL] USBAlternateInterface interface: usbAlternateInterface must inherit property "endpoints" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbAlternateInterface is not defined"
[FAIL] USBEndpoint interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface object length
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface object name
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface: attribute endpointNumber
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface: attribute direction
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface: attribute type
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint interface: attribute packetSize
  assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
[FAIL] USBEndpoint must be primary interface of usbEndpoint
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbEndpoint is not defined"
[FAIL] Stringification of usbEndpoint
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbEndpoint is not defined"
[FAIL] USBEndpoint interface: usbEndpoint must inherit property "endpointNumber" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbEndpoint is not defined"
[FAIL] USBEndpoint interface: usbEndpoint must inherit property "direction" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbEndpoint is not defined"
[FAIL] USBEndpoint interface: usbEndpoint must inherit property "type" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbEndpoint is not defined"
[FAIL] USBEndpoint interface: usbEndpoint must inherit property "packetSize" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: usbEndpoint is not defined"
[FAIL] USBPermissionResult interface: existence and properties of interface object
  assert_own_property: self does not have own property "USBPermissionResult" expected property "USBPermissionResult" missing
[FAIL] USBPermissionResult interface object length
  assert_own_property: self does not have own property "USBPermissionResult" expected property "USBPermissionResult" missing
[FAIL] USBPermissionResult interface object name
  assert_own_property: self does not have own property "USBPermissionResult" expected property "USBPermissionResult" missing
[FAIL] USBPermissionResult interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "USBPermissionResult" expected property "USBPermissionResult" missing
[FAIL] USBPermissionResult interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "USBPermissionResult" expected property "USBPermissionResult" missing
[FAIL] USBPermissionResult interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "USBPermissionResult" expected property "USBPermissionResult" missing
[FAIL] USBPermissionResult interface: attribute devices
  assert_own_property: self does not have own property "USBPermissionResult" expected property "USBPermissionResult" missing
[FAIL] WorkerNavigator interface: attribute usb
  assert_true: The prototype object must have a property "usb" expected true got false
[FAIL] WorkerNavigator interface: navigator must inherit property "usb" with the proper type
  assert_inherits: property "usb" not found in prototype chain
Harness: the test ran to completion.
